Version 1.2 开始搞缓存歌单

+++ HOA-common/src/main/java/context/UserContext.java
@@ -8,7 +8,7 @@ public class UserContext {
         threadLocal.set(id);
     }
 
-    public static Long getCurrentId() {
+    public static Integer getCurrentId() {
         return threadLocal.get();
     }
